[Feat] : 카메라, 플레이어 이동 구현

+++ Risk of Rain 2/Assets/3.Script/ScriptableObjectScript/SurvivorsData.cs
@@ -13,7 +13,7 @@ public class SurvivorsData : ScriptableObject
     public float MoveSpeed;
     public float Mass;
     public float CriticalChance;
-    public float MaxJumpCount;
+    public int MaxJumpCount;
 
     //레벨에 따른 증가량들
